+++ apps/console-electron/app/favicon.ico
