+++ apps/console-electron/package.json
@@ -22,7 +22,6 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "@typescript-eslint/parser": "^7.1.0",
     "concurrently": "^9.0.1",
     "electron": "^32.2.0",
     "eslint": "^8",
