--- apps/console-electron/src/pages/index.tsx
@@ -1,32 +1,0 @@
-import localFont from 'next/font/local';
-import Link from 'next/link';
-
-const geistSans = localFont({
-  src: './fonts/GeistVF.woff',
-  variable: '--font-geist-sans',
-  weight: '100 900',
-});
-const geistMono = localFont({
-  src: './fonts/GeistMonoVF.woff',
-  variable: '--font-geist-mono',
-  weight: '100 900',
-});
-
-export default function Home() {
-  return (
-    <div
-      className={`${geistSans.variable} ${geistMono.variable} grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]`}
-    >
-      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-        <h1 className="lg">DiceDB</h1>
-        <img
-          width={48}
-          height={48}
-          src="https://avatars.githubusercontent.com/u/112580013?s=48&v=4"
-          alt="logo"
-        />
-        <Link href="/dashboard">Home</Link>
-      </main>
-    </div>
-  );
-}
